Create rustfmt_config crate
